model.go: split out record_items.go, cdc_record_stream.go
model.go should mostly contain type definitions
these two types were taking up two thirds of the file with their methods
